+++ src/audio.h
@@ -240,15 +240,17 @@ struct audio_s {
 
 	        ptt_method_t ptt_method; /* none, serial port, GPIO, LPT, HAMLIB, CM108. */
 
-	        char ptt_device[100];	/* Serial device name for PTT.  e.g. COM1 or /dev/ttyS0 */
+	        char ptt_device[128];	/* Serial device name for PTT.  e.g. COM1 or /dev/ttyS0 */
 					/* Also used for HAMLIB.  Could be host:port when model is 1. */
 					/* For years, 20 characters was plenty then we start getting extreme names like this: */
 					/* /dev/serial/by-id/usb-FTDI_Navigator__CAT___2nd_PTT__00000000-if00-port0 */
 					/* /dev/serial/by-id/usb-Prolific_Technology_Inc._USB-Serial_Controller_D-if00-port0 */
 					/* Issue 104, changed to 100 bytes in version 1.5. */
 
-					/* This same field is also used for CM108 GPIO PTT which will */
-					/* have a name like /dev/hidraw1. */
+					/* This same field is also used for CM108/CM119 GPIO PTT which will */
+					/* have a name like /dev/hidraw1 for Linux or */
+					/* \\?\hid#vid_0d8c&pid_0008&mi_03#8&39d3555&0&0000#{4d1e55b2-f16f-11cf-88cb-001111000030} */
+					/* for Windows.  Largest observed was 95 but add some extra to be safe. */
 			
 	        ptt_line_t ptt_line;	/* Control line when using serial port. PTT_LINE_RTS, PTT_LINE_DTR. */
 	        ptt_line_t ptt_line2;	/* Optional second one:  PTT_LINE_NONE when not used. */
@@ -256,7 +258,7 @@ struct audio_s {
 	        int out_gpio_num;	/* GPIO number.  Originally this was only for PTT. */
 					/* It is now more general. */
 					/* octrl array is indexed by PTT, DCD, or CONnected indicator. */
-					/* For CM108, this should be in range of 1-8. */
+					/* For CM108/CM119, this should be in range of 1-8. */
 
 #define MAX_GPIO_NAME_LEN 20	// 12 would cover any case I've seen so this should be safe